--- fastcrypto-tbls/src/weight-reduction/src/solver/mod.rs
@@ -1,4 +1,0 @@
-mod util;
-
-pub mod super_swiper;
-
