--- fastcrypto-tbls/src/weight-reduction/src/solver/util/swiper_common/mod.rs
@@ -1,13 +1,0 @@
-mod bound;
-mod deltas;
-#[cfg(test)]
-mod deltas_test;
-#[cfg(test)]
-mod tests;
-mod tickets;
-
-pub use bound::bound;
-pub use deltas::generate as generate_deltas;
-#[cfg(test)]
-pub(crate) use tests::tests;
-pub use tickets::Tickets;
